Initial commit.
At this point the rendering and caching of multiple variations of a single page
works, as well as the <r:if_session_var> tag and the set_session_vars hook
Page.process.  It also works fine with the memcached gem for the cache, and
probably another cache class with the requisite interface.